Add files via upload
The the grade tracker and the goal suggestion are working perfectly. I need to do the GPS tracker. Also there really isn't many restrictions on the inputs yet. Only the class day must be a valid day and you can't create two of the same courses or two of the same classes or assessments for a course.