Add files via upload
the input on the add classes and add assessments pop ups is different and the addclasses, addassessments, and addgrades functions are different. I'm can't figure out why it's giving an undefined output where it should output the month in the next 7 days pop up. I figured out how to create drop down lists for the courses. I'm working on that now, so you don't have to enter it manually.